avoid truncation of zoom factor used for DPI scaling
This was originally reported and already fixed upstream:
https://bugs.webkit.org/show_bug.cgi?id=40662

However, when porting c1af97f5f73477f62427bdf0b6e1480b6f75b6f3 to
the latest Qt version this fix was lost (due to code moving around).

wkhtmltopdf/qt#12

+++ src/3rdparty/webkit/Source/WebKit/qt/Api/qwebframe.cpp
@@ -223,8 +223,8 @@ QWebPrinterPrivate::QWebPrinterPrivate(const QWebFrame *f, QPaintDevice *printer
     , frame(f)
     , graphicsContext(&p)
 {
-    const qreal zoomFactorX = printer->logicalDpiX() / qt_defaultDpi();
-    const qreal zoomFactorY = printer->logicalDpiY() / qt_defaultDpi();
+    const qreal zoomFactorX = (qreal)printer->logicalDpiX() / qt_defaultDpi();
+    const qreal zoomFactorY = (qreal)printer->logicalDpiY() / qt_defaultDpi();
     IntRect pageRect(0, 0,
                      int(printer->width() / zoomFactorX),
                      int(printer->height() / zoomFactorY));
